feat: RecordService 게임플레이횟수

+++ src/main/java/com/example/marimo_back/Dto/RecordResponseDto.java
@@ -23,7 +23,7 @@ public class RecordResponseDto {
 
     private List<String> mostFail;
 
-    private Integer gameJoinNum;
+    private Long gameJoinNum;
 
-    private Integer taleJoinNum;
+    private Long taleJoinNum;
 }
